fix(cf-deploy-config-writer): cleanup code
