+++ packages/cf-deploy-config-writer/src/utils.ts
@@ -36,6 +36,9 @@ import {
 } from './constants';
 import type { Editor } from 'mem-fs-editor';
 import { type MTABaseConfig, type CFConfig, type CFBaseConfig, type CFAppConfig } from './types';
+import LoggerHelper from './logger-helper';
+import { CommandRunner } from '@sap-ux/nodejs-utils';
+import { t } from './i18n';
 
 let cachedDestinationsList: Destinations = {};
 
@@ -295,4 +298,42 @@ export async function alignCdsVersions(rootPath: string, fs: Editor): Promise<vo
     if (!cdsDKDevDepVersion && cdsDepVersion) {
         await addPackageDevDependency(rootPath, CDSDKPackage, cdsDepVersion, fs);
     }
+}
+
+/**
+ * Executes a command in the specified project directory.
+ *
+ * @async
+ * @param {string} cwd - Working directory where the command will be executed
+ * @param {string} cmd - Command to execute
+ * @param {string[]} args - Arguments to pass to the command
+ * @param {string} errorMsg - Error message prefix to display if the command fails
+ * @returns {Promise<void>} - A promise that resolves when the command completes successfully
+ * @throws {Error} Throws an error with the provided error message concatenated with the original error if execution fails
+ * @example
+ * // Execute npm install in the project directory
+ * await runCommand('/path/to/project', 'npm', ['install'], 'Failed to install dependencies:');
+ */
+export async function runCommand(cwd: string, cmd: string, args: string[], errorMsg: string): Promise<void> {
+    const commandRunner = new CommandRunner();
+    try {
+        await commandRunner.run(cmd, args, { cwd });
+    } catch (e) {
+        throw new Error(`${errorMsg} ${e.message}`);
+    }
+}
+
+/**
+ * Run npm install in the specified project directory.
+ *
+ * @param rootPath Path to the project directory
+ * @param fs reference to a mem-fs editor
+ */
+export async function runNpmInstall(rootPath: string, fs: Editor): Promise<void> {
+    // When installing, we need to ensure that the package.json is written to disk before running npm install
+    await new Promise((resolve) => fs.commit(resolve));
+    LoggerHelper?.logger?.debug(`npm install command will be executed in ${rootPath}`);
+    const cmd = process.platform === 'win32' ? `npm.cmd` : 'npm';
+    // Install latest dev dependencies, if any, added by the CF writer
+    await runCommand(rootPath, cmd, ['install', '--ignore-engines'], t('error.errorInstallingNodeModules'));
 }
